fix(correction pt.2)

+++ Library.Services/Implementations/BookUseCases/DeleteBookUseCase.cs
@@ -1,6 +1,7 @@
 ﻿using Library.Application.Interfaces.Repositories;
 using Library.Application.Interfaces.Services;
-using Library.Domain.Interfaces.BookUseCases;
+using Library.Application.Interfaces.BookUseCases;
+using Library.Application.Exceptions;
 
 namespace Library.Application.Implementations.BookUseCases {
     
@@ -16,6 +17,9 @@ namespace Library.Application.Implementations.BookUseCases {
         public async Task Execute( Guid bookId ) {
             _unit.CreateTransaction();
             var book = await _unit.bookRepository.GetAsync( bookId );
+            if (book == null) {
+                throw new NotFoundException( "Такой книги не существует!" );
+            }
             await _unit.bookRepository.DeleteAsync( book );
             _imageService.DeleteImage( bookId ); 
             await _unit.Save();
